requirements.txt
awslambdaric
boto3
pennsieve2
requests
python-dotenv
